+++ tests/testthat/test-vroom_lines.R
@@ -31,3 +31,18 @@ test_that("vroom_lines works with connections files", {
 
   expect_equal(actual, expected)
 })
+
+
+test_that("vroom_lines works with files with no trailing newline", {
+  f <- tempfile()
+  on.exit(unlink(f))
+
+  writeBin(charToRaw("foo"), f)
+  expect_equal(vroom_lines(f), "foo")
+
+  f2 <- tempfile()
+  on.exit(unlink(f2), add = TRUE)
+
+  writeBin(charToRaw("foo\nbar"), f2)
+  expect_equal(vroom_lines(f2), c("foo", "bar"))
+})